+++ src/gspn/gspn_io.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "common.h"
+
+void define_gspn_io(py::module& m);
+