+++ src/mod_pomdp.cpp
@@ -4,6 +4,8 @@
 #include "pomdp/tracker.h"
 #include "pomdp/qualitative_analysis.h"
 #include "pomdp/transformations.h"
+#include "pomdp/memory.h"
+#include <storm/adapters/RationalFunctionAdapter.h>
 
 PYBIND11_MODULE(pomdp, m) {
     m.doc() = "Functionality for POMDP analysis";
@@ -15,5 +17,8 @@ PYBIND11_MODULE(pomdp, m) {
     define_tracker(m);
     define_qualitative_policy_search<double>(m, "Double");
     define_qualitative_policy_search_nt(m);
+    define_memory(m);
+    define_transformations_nt(m);
     define_transformations<double>(m, "Double");
+    define_transformations<storm::RationalFunction>(m, "Rf");
 }