+++ src/storage/dd.cpp
@@ -1,5 +1,6 @@
 #include "dd.h"
 #include "storm/storage/dd/DdManager.h"
+#include "storm/storage/dd/DdMetaVariable.h"
 #include "storm/storage/dd/Dd.h"
 #include "storm/storage/dd/Bdd.h"
 #include "src/helpers.h"
@@ -9,9 +10,11 @@ void define_dd(py::module& m, std::string const& libstring) {
     py::class_<storm::dd::DdMetaVariable<DdType>> ddMetaVariable(m, (std::string("DdMetaVariable_") + libstring).c_str());
     ddMetaVariable.def("compute_indices", &storm::dd::DdMetaVariable<DdType>::getIndices, py::arg("sorted")=true);
     ddMetaVariable.def_property_readonly("name", &storm::dd::DdMetaVariable<DdType>::getName);
+    ddMetaVariable.def_property_readonly("lowest_value", &storm::dd::DdMetaVariable<DdType>::getLow);
+    ddMetaVariable.def("__str__", &storm::dd::DdMetaVariable<DdType>::getName);
 
     py::class_<storm::dd::DdManager<DdType>, std::shared_ptr<storm::dd::DdManager<DdType>>> ddManager(m, (std::string("DdManager_") + libstring).c_str());
-    //ddManager.def("get_meta_variable", &storm::dd::DdManager<DdType>::getMetaVariable, py::arg("expression_variable"));
+    ddManager.def("get_meta_variable", [](storm::dd::DdManager<DdType> const& manager, storm::expressions::Variable const& var) {return manager.getMetaVariable(var);}, py::arg("expression_variable"));
 
     py::class_<storm::dd::Dd<DdType>> dd(m, (std::string("Dd_") + libstring).c_str(), "Dd");
     dd.def_property_readonly("node_count", &storm::dd::Dd<DdType>::getNodeCount, "get node count");